requirements.txt
requests
setuptools
pyyaml
python-dotenv
pandas
numpy
Faker
SQLAlchemy
sqlalchemy-trino
pystarburst
openai
pydantic
google-generativeai
gradio
-e .